--- static/app/components/core/asyncCompactSelect/index.tsx
@@ -1,1 +1,0 @@
-export {AsyncCompactSelect} from './asyncCompactSelect';
